Added support for external outbound hostname in the proxy config.
The outbound hostname feature is useful when the proxy is not part of the
swarm. In a large corporate network setting, swarms may be regularly
decommissioned and re-created, but external parts integrating with the
swarms can not be reconfigured. (Updating DNS records is often times out
of the question for policy and/or bureaucracy reasons.)