Change default SERVICE_DOMAIN_ALGO to hdr_beg. (#360)
if published port is something other than 80 or 443, browsers will include the port in the Requests Header Host attribute.  If HAProxy's ACL is checking for hdr(host) to be some domain, it will not match.  Instead, we will use hdr_beg(host) to only look at the begging of the Host value for a match.