Issue 364 (#391)
* Switched over to using a buffer for stdout,stderr on execute #364 (#367)

Instead of using Combinedout, switched to using a pipe for stdout and stderr. It seems like running a command on HAProxy doesn't capture the entire stdout and stderr. This will ensure we get everything and see what is actually going if there is an error.

* Fixes #364

* Fixes #364

* Fixed #336

* Fix exit status on verify haproxy config

This commit will check for an exit status on verifying the haproxy.cfg. HAproxy may output warnings as an error, but overall it would be considered a valid text and wouldn't stop HAproxy from running.

* Removed obsolete test for pr #384