Override health check port with $PORT
The proxy server allows you to configure the port that it runs on with the environment variable `PORT`. If you do that, the health check will fail because 8080 is hard-coded and docker-flow-proxy will continually reload itself. This will use the port from `PORT` and fallback to 8080 if it's not found